+++ src/pages/components/Home.jsx
@@ -7,12 +7,11 @@ import Tabs from '@mui/material/Tabs';
 import Tab from '@mui/material/Tab';
 import Typography from '@mui/material/Typography';
 import Box from '@mui/material/Box';
-import Content from './question/Content';
+import PostQuestion from './question/PostQuestion';
 import IndexQuestion from './question/questionList/IndexQuestion';
 
 import { Link  as LinkRouter } from 'react-router-dom';
 import QuestionPage from './question/questionList/ShowQuestion'
-import QuestionUser from './question/QuestionUser'
 
 function TabPanel(props) {
   const { children, value, index, ...other } = props;
@@ -81,7 +80,7 @@ export default function Home() {
       >
         {/* 質問 */}
         <TabPanel value={value} index={0} dir={theme.direction} >
-          <Content />
+          <PostQuestion />
         </TabPanel>
 
         {/* 質問一覧 */}
